refactor: extract reusable DemoStoreSeeder from V006 local seed
